+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/security/jwt/RefreshTokenRedisRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, Long> {
+public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, String> {
     Optional<RefreshToken> findByRefreshToken(String refreshToken);
 }
