+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/security/jwt/filter/JwtAuthenticationFilter.java
@@ -22,7 +22,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+import static kr.ac.kumoh.illdang100.tovalley.util.CookieUtil.addCookie;
 import static kr.ac.kumoh.illdang100.tovalley.util.CustomResponseUtil.*;
+import static kr.ac.kumoh.illdang100.tovalley.util.HttpServletUtil.getClientIpAddress;
+import static kr.ac.kumoh.illdang100.tovalley.util.TokenUtil.*;
 
 @Slf4j
 public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
