+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/security/jwt/filter/JwtAuthorizationFilter.java
@@ -21,8 +21,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+import static kr.ac.kumoh.illdang100.tovalley.util.CookieUtil.*;
 import static kr.ac.kumoh.illdang100.tovalley.util.CustomResponseUtil.*;
-import static kr.ac.kumoh.illdang100.tovalley.util.CustomResponseUtil.addCookie;
 import static kr.ac.kumoh.illdang100.tovalley.util.EntityFinder.findRefreshTokenOrElseThrowEx;
 
 @Slf4j
