+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/security/oauth/OAuth2FailureHandler.java
@@ -17,7 +17,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-import static kr.ac.kumoh.illdang100.tovalley.util.CustomResponseUtil.addCookie;
+import static kr.ac.kumoh.illdang100.tovalley.util.CookieUtil.addCookie;
 
 @Slf4j
 @RequiredArgsConstructor
