+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/security/oauth/OAuth2SuccessHandler.java
@@ -16,7 +16,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+import static kr.ac.kumoh.illdang100.tovalley.util.CookieUtil.addCookie;
 import static kr.ac.kumoh.illdang100.tovalley.util.CustomResponseUtil.*;
+import static kr.ac.kumoh.illdang100.tovalley.util.HttpServletUtil.getClientIpAddress;
+import static kr.ac.kumoh.illdang100.tovalley.util.TokenUtil.*;
 
 @Slf4j
 @RequiredArgsConstructor
