+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/util/CustomResponseUtil.java
@@ -1,23 +1,12 @@
 package kr.ac.kumoh.illdang100.tovalley.util;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import kr.ac.kumoh.illdang100.tovalley.domain.member.Member;
 import kr.ac.kumoh.illdang100.tovalley.dto.ResponseDto;
-import kr.ac.kumoh.illdang100.tovalley.security.jwt.JwtProcess;
-import kr.ac.kumoh.illdang100.tovalley.security.jwt.JwtVO;
-import kr.ac.kumoh.illdang100.tovalley.security.jwt.RefreshToken;
-import kr.ac.kumoh.illdang100.tovalley.security.jwt.RefreshTokenRedisRepository;
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.net.URLDecoder;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
 
 public class CustomResponseUtil {
 
@@ -48,91 +37,6 @@ public class CustomResponseUtil {
         } catch (Exception e) {
             log.error("서버 파싱 에러");
         }
-    }
-
-    public static void addCookie(HttpServletResponse response, String name, String value, boolean httpOnly) {
-        value = URLEncoder.encode(value, StandardCharsets.UTF_8);
-        Cookie cookie = new Cookie(name, value);
-        cookie.setHttpOnly(httpOnly);
-        cookie.setPath("/");
-        response.addCookie(cookie);
-    }
-
-    public static void addCookie(HttpServletResponse response, String name, String value) {
-        addCookie(response, name, value, true);
-    }
-
-    public static String saveRefreshToken(JwtProcess jwtProcess, RefreshTokenRedisRepository refreshTokenRedisRepository, Member member, String ip) {
-        String memberId = member.getId().toString();
-        String role = member.getRole().toString();
-
-        String refreshToken = jwtProcess.createRefreshToken(memberId, role, ip);
-
-        refreshTokenRedisRepository.save(RefreshToken.builder()
-                .id(memberId)
-                .role(role)
-                .ip(ip)
-                .refreshToken(refreshToken)
-                .build());
-
-        return refreshToken;
-    }
-
-
-    public static boolean isCookieVerify(HttpServletRequest request, String cookieName) {
-        Cookie[] cookies = request.getCookies();
-        if (cookies != null) {
-            for (Cookie cookie : cookies) {
-                String cookieValue = URLDecoder.decode(cookie.getValue(), StandardCharsets.UTF_8);
-                if (cookie.getName().equals(cookieName) && cookieValue.startsWith(JwtVO.TOKEN_PREFIX)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    public static String findCookieValue(HttpServletRequest request, String cookieName) {
-        String token = "";
-        Cookie[] cookies = request.getCookies();
-        for (Cookie cookie : cookies) {
-            String cookieValue = URLDecoder.decode(cookie.getValue(), StandardCharsets.UTF_8);
-            if (cookie.getName().equals(cookieName) && cookieValue.startsWith(JwtVO.TOKEN_PREFIX)) {
-                token = cookieValue;
-            }
-        }
-        return token;
-    }
-
-    public static String getClientIpAddress(HttpServletRequest request) {
-        String ipAddress = request.getHeader("X-Forwarded-For");
-
-        if (StringUtils.isBlank(ipAddress) || "unknown".equalsIgnoreCase(ipAddress)) {
-            ipAddress = request.getHeader("X-Forwarded-For");
-        }
-        if (StringUtils.isBlank(ipAddress) || "unknown".equalsIgnoreCase(ipAddress)) {
-            ipAddress = request.getHeader("X-Real-IP");
-        }
-        if (StringUtils.isBlank(ipAddress) || "unknown".equalsIgnoreCase(ipAddress)) {
-            ipAddress = request.getHeader("X-Real-IP");
-        }
-        if (StringUtils.isBlank(ipAddress) || "unknown".equalsIgnoreCase(ipAddress)) {
-            ipAddress = request.getHeader("Http_X_FORWARDED_FOR");
-        }
-        if (StringUtils.isBlank(ipAddress) || "unknown".equalsIgnoreCase(ipAddress)) {
-            ipAddress = request.getHeader("HTTP_CLIENT_IP");
-        }
-        if (StringUtils.isBlank(ipAddress) || "unknown".equalsIgnoreCase(ipAddress)) {
-            ipAddress = request.getHeader("Proxy-Client-IP");
-        }
-        if (StringUtils.isBlank(ipAddress) || "unknown".equalsIgnoreCase(ipAddress)) {
-            ipAddress = request.getHeader("WL-Proxy-Client-IP");
-        }
-        if (StringUtils.isBlank(ipAddress) || "unknown".equalsIgnoreCase(ipAddress)) {
-            ipAddress = request.getRemoteAddr();
-        }
-
-        return ipAddress;
     }
 }
